refactor(domain): `TripEstimate#calculateTotalFare` 로직을 `TaxiType` enum 클래스에 위임한다.
