+++ core/core-domain/src/main/java/com/giwankim/taxiservice/core/domain/application/port/in/GetTripEstimatesUseCase.java
@@ -1,7 +1,5 @@
 package com.giwankim.taxiservice.core.domain.application.port.in;
 
-import com.giwankim.taxiservice.core.domain.domain.TripEstimates;
-
 public interface GetTripEstimatesUseCase {
   TripEstimates getTripEstimates(GetTripEstimatesQuery query);
 }
