+++ core/core-domain/src/test/java/com/giwankim/taxiservice/core/domain/domain/TaxiTypeTest.java
@@ -3,41 +3,20 @@ package com.giwankim.taxiservice.core.domain.domain;
 import static com.giwankim.taxiservice.domain.Fixtures.aDirections;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.giwankim.taxiservice.core.domain.domain.surcharge.AmountSurchargePolicy;
-import com.giwankim.taxiservice.core.domain.domain.surcharge.NoneSurchargePolicy;
-import com.giwankim.taxiservice.core.enums.TaxiType;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class TripEstimateTest {
-
-  @ParameterizedTest(name = "{0}")
-  @MethodSource("provideArgumentsForCreate")
-  @DisplayName("create 메서드 테스트")
-  void create(TaxiType taxiType, Class<?> expectedSurchargePolicyClass) {
-    KRWMoney baseFare = KRWMoney.wons(10_000);
-    assertThat(TripEstimate.create(taxiType, baseFare).getSurchargePolicy())
-        .isInstanceOf(expectedSurchargePolicyClass);
-  }
-
-  private static Stream<Arguments> provideArgumentsForCreate() {
-    return Stream.of(
-        Arguments.of(TaxiType.REGULAR, NoneSurchargePolicy.class),
-        Arguments.of(TaxiType.DELUXE, AmountSurchargePolicy.class),
-        Arguments.of(TaxiType.JUMBO, AmountSurchargePolicy.class));
-  }
+class TaxiTypeTest {
 
   @ParameterizedTest(name = "택시 유형: {0}, 기본 요금: {1}, 예상 요금: {3}")
   @MethodSource("provideArgumentsForCalculateTotalFare")
   @DisplayName("calculateTotalFare 메서드 테스트")
   void calculateTotalFare(
       TaxiType taxiType, KRWMoney baseFare, Directions directions, KRWMoney expectedTotalFare) {
-    TripEstimate tripEstimate = TripEstimate.create(taxiType, baseFare);
-    KRWMoney actualTotalFare = tripEstimate.calculateTotalFare(directions);
-    assertThat(actualTotalFare).isEqualTo(expectedTotalFare);
+    assertThat(taxiType.calculateTotalFare(baseFare, directions)).isEqualTo(expectedTotalFare);
   }
 
   private static Stream<Arguments> provideArgumentsForCalculateTotalFare() {
